requirements.txt
annotated-types==0.7.0
anyio==4.11.0
black==25.11.0
brotli==1.2.0
cachetools==6.2.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
ddgs==9.8.0
defusedxml==0.7.1
distro==1.9.0
dotenv==0.9.9
et_xmlfile==2.0.0
fonttools==4.61.1
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.2.1
google-genai==1.57.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.11.1
kiwisolver==1.4.9
lxml==6.0.2
markdown-it-py==4.0.0
matplotlib==3.10.8
mdurl==0.1.2
mypy_extensions==1.1.0
numpy==1.26.4
odf==0.0.1
odfpy==1.4.1
olefile==0.47
openai==2.7.1
openpyxl==3.1.2
packaging==25.0
pandas==2.1.4
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
primp==0.15.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg2==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.2.5
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.2.1
python-pptx==1.0.2
python-telegram-bot==22.5
pytokens==0.3.0
pytz==2025.2
regex==2025.11.3
reportlab==4.4.7
requests==2.32.5
rich==14.2.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
telegram==0.0.1
tenacity==9.1.2
tesseract==0.1.3
tiktoken==0.8.0
tqdm==4.67.1
typer==0.12.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0.1
xlrd==2.0.1
xlsxwriter==3.2.9
